// src/app/interfaces/movie-description.ts
export interface MovieDescription {
           genre?: object;
           actor?: object;
           title?: string;
        premiere?: boolean;
         popular?: boolean;
     description?: string;
          length?: string;
           price?: number;
            year?: number;
          rating?: number;
     mainPopular?: boolean;
    mainPremiere?: boolean;
        mainHome?: boolean;
       numOrders?: number;
      trailerUrl?: string;
          poster?: string;
      background?: string;
}